+++ common/src/main/kotlin/dev/sterner/witchery/registry/WitcheryPayloads.kt
@@ -58,6 +58,9 @@ object WitcheryPayloads {
         registerS2C(SyncVoodooDataS2CPacket.ID, SyncVoodooDataS2CPacket.STREAM_CODEC) { payload, context ->
             payload.handleS2C(payload, context)
         }
+        registerS2C(OpenLecternGuidebook.ID, OpenLecternGuidebook.STREAM_CODEC) { payload, context ->
+            payload.handleS2C(payload, context)
+        }
 
 
         //C2S