+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Bugfix 0.1.2e
 - Removed Infinity Egg from OP tab
+- Fixed Fluid capability on neoforge
+- Fixed wood ash reset
+- Fixed Lectern handling
+- Fixed recipes
+- Fixed interdimensional travel 
 
 ### Bugfix 0.1.2d
 - Fixed Wormwood lang and collision